refactor: unify server fixtures
