+++ package.json
@@ -86,7 +86,6 @@
     "eslint": "^7.23.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^3.3.1",
-    "get-port": "^5.1.1",
     "glob": "^7.1.6",
     "graphql": "^15.5.0",
     "jest": "^26.6.3",
