+++ src/tests/use/ws.ts
@@ -6,7 +6,7 @@ import {
   parseMessage,
   SubscribePayload,
 } from '../../message';
-import { startTServer } from '../fixtures/simple';
+import { startWSTServer as startTServer } from '../fixtures/simple';
 import { createTClient } from '../utils';
 
 it('should gracefully go away when disposing', async () => {
